logaritma desimal

+++ body.c
@@ -408,7 +408,9 @@ void convertPostfix(Queue *Z,Stack *X,char *input){
 				b=DequeOperand(&*Z);
 				basis_bebas=1;
 			}
-			float hasil3=0,hasil2;
+
+			double hasil3=0,hasil2;
+
 			int o=1,t=1;
 			char valid[5];
 			valid[4]='\0';
